finished upgrade implemtation
-added upgrades to all weapons
-added actions to all weapons
-had to touch a lot of code
-made test suit in anthony volume test scene

+++ Assets/Programming/Scripts/Weapon/Weapon.cs
@@ -78,7 +78,7 @@ public abstract class Weapon : Tool
 public struct WeaponScalars
 {
     public float Damage;
-    public int ClipSize;
+    public float ClipSize;
     public float AmmoReserveSize;
     public float ReloadTime;
     public float FireRate;
